Gradient-based climate velocity

Function to calculate the velocity of climate change after Burrows et al. (2011) based on local climatic temporal trends and spatial gradients.

Usage

gVoCC(tempTrend, spatGrad)

Arguments

tempTrend

The output from the tempTrend function containing the long-term linear climatic trends.

spatGrad

The output from the spatGrad function containing the magnitudes and angles for the spatial climatic gradient.

Value

A RasterStack containing the climate velocity magnitude ("voccMag", km/yr for unprojected rasters and spatial unit/year for projected rasters) and angle("voccAng" in degrees north: 0N, 90E, 180S and 270W).

References

Burrows et al. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science, 334, 652-655.

See also

Author

Jorge Garcia Molinos

Examples

if (FALSE) { # \dontrun{
HSST <- VoCC_get_data("HSST.tif")
yrSST <- sumSeries(HSST,
  p = "1960-01/2009-12", yr0 = "1955-01-01", l = terra::nlyr(HSST),
  fun = function(x) colMeans(x, na.rm = TRUE), freqin = "months", freqout = "years"
)
tr <- tempTrend(yrSST, th = 10)
sg <- spatGrad(yrSST, th = 0.0001, projected = FALSE)

# Magnitude and angle of the climate velocity (km/yr) 1960-2009

v <- gVoCC(tr, sg)
terra::plot(v)
} # }