Shift in timing of seasonal climatology

Function to calculate the seasonal shift in the arriving of typical seasonal climates for a given period of interest as per Burrows et al. (2011).

Usage

shiftTime(r, yr1, yr2, yr0, th, m)

Arguments

r

stack with monthly values of the climatic variable for the period of interest.

yr1

integer specifying the initial year for the period of interest.

yr2

integer specifying the end year for the period of interest.

yr0

integer specifying the first year in the series.

th

integer minimum number of non NAs in the series needed to calculate the trend (default 3).

m

integer number (1-12) of the month for which the shift is to be calculated

Value

a stack with the long-term monthly trend (C/year for temperature in degrees; "mTrend"), seasonal rate of change (C/month; "seaRate"), and seasonal shift (day/decade; "seaShift").

References

Burrows et al. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science, 334, 652-655.

Author

Jorge Garcia Molinos and Michael T. Burrows

Examples

if (FALSE) { # \dontrun{
HSST <- VoCC_get_data("HSST.tif")
Apr <- shiftTime(HSST, yr1 = 1960, yr2 = 2009, yr0 = 1955, th = 10, m = 4)

terra::plot(Apr)
} # }